+++ src/core/namespace.js
@@ -3,13 +3,13 @@
 
   const family = globalThis.ToolboxFamily || {};
   family.Core ||= {};
-  family.version = '1.0.0-rc.3';
+  family.version = '1.0.0-rc.4';
 
   const toolbox = globalThis.RedditToolbox || {};
   toolbox.Core = family.Core;
   toolbox.Reddit ||= {};
   toolbox.UI ||= {};
-  toolbox.version = '1.0.0-rc.3';
+  toolbox.version = '1.0.0-rc.4';
 
   globalThis.ToolboxFamily = family;
   globalThis.RedditToolbox = toolbox;